Updated filtering of output resources based on a global property.
Removed dependency on jdt.
Added filtering of resources to also to delta visitors.

The filtering path is global for the moment, and it accepts only one
project relative path. In the future it will be moved to a project
property and will accept more than one path (or path pattern).